backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
qdrant-client>=1.7.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
requests>=2.31.0
PyPDF2>=3.0.0
docx2txt>=0.8
python-multipart>=0.0.6
